fix(ci): line-count guardrail — decompose github_client & bitbucket_client under 1500 lines
test_first_party_code_files_stay_under_line_limit failed because retry-policy
work pushed both common/github_client.py (1511) and common/bitbucket_client.py
(1532) past the 1500-line maintainability limit. Extract cohesive pure code into
existing/adjacent sibling modules (the established decomposition pattern) with no
behavior change:

- Move the module-level rollup/branch-rule helpers and their two constants from
  github_client.py into github_client_parsing.py, re-importing what the client
  still uses (github_client 1511 -> 1464).
- Move the stateless Bitbucket REST-path/web-URL builders into a new
  bitbucket_client_paths._BitbucketUrlsMixin that BitbucketClient inherits, so
  every self._repo_path(...) call site is unchanged (bitbucket_client 1532 -> 1499).
